Misc fixes

+++ Hazel/src/Hazel/Renderer/Buffer.h
@@ -41,7 +41,7 @@ namespace Hazel
     	BufferElement(const ShaderDataType type, const std::string& name, const bool normalized = false)
     		: Name(name), Type(type), Size(ShaderDataTypeSize(type)), Offset(0), Normalized(normalized) { }
 
-    	uint32_t GetComponentCount() const
+    	uint8_t GetComponentCount() const
     	{
     		switch (Type)
     		{
